Changing color animation demonstration.
Originally, all 3 animations used red, green, and blue sprites. I think
they were even in the same order. This made it practically impossible to
see when the animation changed (aside from the corresponding frame rate
change). To fix that, I've shortened the animations to be two colors
instead of three, which allows you to see the animation change more
easily.