h4x: Track when the player has "landed".
This is useful for determining whether or not he can run[1] or jump, but
it's also a hack since it isn't really taking true collision detection
into consideration. I'm not sure how this should be modelled yet. This
is the quick and dirty way to get it semi-functional for Santahack
(albeit, time has run out).